build: resolve a -Wconversion compiler warning
wordEncoder.cpp:312:91: warning: conversion from
‘std::__cxx11::basic_string<char>::size_type’ {aka ‘long unsigned
int’} to ‘double’ may change value [-Wconversion]
  312 | buffer.length() == 0 ? 1 : static_cast<double>(asciiCount) / buffer.length();
      |                                                              ~~~~~~~~~~~~~^~

+++ src/vmime/wordEncoder.cpp
@@ -309,7 +309,7 @@ wordEncoder::Encoding wordEncoder::guessBestEncoding(
 		utility::stringUtils::countASCIIchars(buffer.begin(), buffer.end());
 
 	const double asciiPercent =
-		buffer.length() == 0 ? 100 : static_cast<double>(asciiCount) / buffer.length();
+		buffer.length() == 0 ? 100 : static_cast<double>(asciiCount) / static_cast<double>(buffer.length());
 
 	if (asciiPercent < 0.60) {
 		return ENCODING_B64;
